+++ feature/home/src/main/java/com/linku/home/screen/AlarmScreen.kt
@@ -1,21 +1,13 @@
 package com.linku.home.screen
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -24,27 +16,20 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.paging.LoadState
 import androidx.paging.PagingData
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
-import androidx.paging.compose.itemKey
-import com.linku.core.error.AppError
 import com.linku.core.model.alarm.AlarmSummary
 import com.linku.core.model.alarm.AlarmType
-import com.linku.design.modifier.skeleton
 import com.linku.design.theme.LinkuPreview
 import com.linku.design.theme.LocalColorTheme
-import com.linku.home.ui.alarm.component.AlarmItem
-import com.linku.home.ui.alarm.component.AlarmNothingTab
 import com.linku.home.ui.alarm.component.AlarmTopBar
 import com.linku.home.ui.alarm.component.AlarmFilterTabs
-import com.linku.home.ui.alarm.component.AlarmListContent
+import com.linku.home.ui.alarm.component.listcontent.AlarmListContent
 import com.linku.home.ui.alarm.component.AlarmSettingTab
 import com.linku.home.viewmodel.AlarmViewModel
 import kotlinx.coroutines.flow.flowOf
@@ -117,6 +102,7 @@ private fun AlarmScreenContent(
             onClick = onNavigateToMyPage
         )
 
+        // 허용이 되어있는 경우
         if (isAlarmAllowed) {
             AlarmListContent(
                 alarms = alarms,
